% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional_mean_comparison.R
\name{estimate_nuisance_parameter_lasso}
\alias{estimate_nuisance_parameter_lasso}
\title{The function for nuisance parameter estimation in anchored_lasso_testing().}
\usage{
estimate_nuisance_parameter_lasso(
  nuisance_sample_1,
  nuisance_sample_2,
  pca_method = "sparse_pca",
  mean_method = "lasso",
  num_latent_factor = 1,
  local_environment = local_environment
)
}
\arguments{
\item{nuisance_sample_1}{Group 1 sample. Each row is a subject and each column corresponds to a feature.}

\item{nuisance_sample_2}{Group 2 sample. Each row is a subject and each column corresponds to a feature.}

\item{pca_method}{Methods used to estimate principle component The default is "sparse_pca", using sparse PCA from package PMA. Other choices are "dense_pca"---the regular PCA; and "hard"--- hard-thresholding PCA, which also induces sparsity.}

\item{mean_method}{Methods used to estimate the discriminant direction. Default is logistic Lasso "lasso".}

\item{num_latent_factor}{The principle component that lasso coefficient anchors at. The default is PC1 = 1.}

\item{local_environment}{A environment for hyperparameters shared between folds.}
}
\description{
The function for nuisance parameter estimation in anchored_lasso_testing().
}
