% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional_mean_comparison.R
\name{evaluate_pca_plug_in}
\alias{evaluate_pca_plug_in}
\title{Calculate the test statistics on the left-out samples. Called in simple_pc_testing().}
\usage{
evaluate_pca_plug_in(
  cross_fitting_sample_1,
  cross_fitting_sample_2 = NULL,
  nuisance_collection
)
}
\arguments{
\item{cross_fitting_sample_1}{Group 1 sample. Each row is a subject and each column corresponds to a feature.}

\item{cross_fitting_sample_2}{Group 2 sample. Each row is a subject and each column corresponds to a feature.}

\item{nuisance_collection}{A collection of nuisance quantities estimated using "nuisance" samples. It is the output of estimate_nuisance_pc().}
}
\description{
Calculate the test statistics on the left-out samples. Called in simple_pc_testing().
}
