% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_function.R
\name{extract_lasso_coef}
\alias{extract_lasso_coef}
\title{Extract the lasso estimate from the output of anchored_lasso_testing().}
\usage{
extract_lasso_coef(testing_result)
}
\arguments{
\item{testing_result}{The output/test result list from anchored_lasso_testing().}
}
\value{
A list, whose elements are the estimated discriminant directions for each split---the length of the output list is the same as n_folds.
}
\description{
Extract the lasso estimate from the output of anchored_lasso_testing().
}
