% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_pre_processing_functions.R
\name{index_spliter}
\alias{index_spliter}
\title{Split indices into folds}
\usage{
index_spliter(array, n_folds = 5)
}
\arguments{
\item{array}{A vector of indices (e.g., \code{1:n}) to be split into folds.}

\item{n_folds}{Integer. Number of folds. Default is 5.}
}
\value{
A list of length \code{n_folds}, each containing a subset of the shuffled indices.
}
\description{
Randomly splits a given vector of indices into approximately equal-sized folds.
}
\examples{
index_spliter(1:10, n_folds = 3)

}
